feat: Comprehensive project improvements
This commit introduces a wide range of improvements to the YINI project, based on a comprehensive review of the codebase.

The key improvements include:

-   **Unified Build System:** The C++ and C# build processes have been consolidated into a single CMake build system, and the `build.py` script has been removed. This simplifies the build process and makes it more maintainable.
-   **Enhanced C-API:** The C-API has been made more robust with the addition of a detailed error-handling mechanism. A new `yini_manager_get_last_error` function provides descriptive error messages, and a `yini_manager_has_key` function allows for efficient key existence checks.
-   **Improved C# Wrapper:** The C# wrapper has been updated to leverage the new C-API features. It now uses a custom `YiniException` for clearer error reporting and includes a `HasKey` method for better performance.
-   **Code Quality:** Numerous compiler warnings in the C++ codebase have been fixed, improving code quality and stability.
-   **Updated Tests:** The C# test suite has been updated to align with the new exception-based error handling, ensuring the tests remain accurate and comprehensive.

These changes make the YINI project more robust, user-friendly, and maintainable.

+++ src/Lexer/Lexer.cpp
@@ -110,13 +110,13 @@ namespace YINI
     char Lexer::peek()
     {
         if (isAtEnd()) return '\0';
-        return m_source[m_current];
+        return m_source[static_cast<size_t>(m_current)];
     }
 
     char Lexer::peekNext()
     {
-        if (m_current + 1 >= m_source.length()) return '\0';
-        return m_source[m_current + 1];
+        if (static_cast<size_t>(m_current + 1) >= m_source.length()) return '\0';
+        return m_source[static_cast<size_t>(m_current + 1)];
     }
 
     char Lexer::advance()
